A getting started video will help Windows users.
Although linux and mac users are likely to already know how to get started with Docker, a better explanation was needed for Windows users who may never have used bash or Docker. A short video was produced and is hosted on Fidelity's video site. This commit adds instructions, image and link to the README.